feat: add user slice
